Updating script.

+++ install.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 
 if [ -z "$USER" ]; then
     USER=$(id -un)
@@ -13,9 +13,9 @@ echo >&2 "====================================================================="
 
 # Install the dotfiles I want
 cp -r dotfiles.codespaces $HOME/.dotfiles.codespaces
-if [[ -z "$PAT_TOKEN" ]]
+if [[ -z "$PAT_TOKEN" ]] #this
 then
-  git clone --depth 1 --recurse-submodules --shallow-submodules https://${GITHUB_USER}:${PAT_TOKEN}@github.com/${GITHUB_USER}/dotfiles.git $HOME/.dotfiles
+  git clone --depth 1 --recurse-submodules --shallow-submodules https://${GITHUB_USER}:${PAT_TOKEN}@github.com/${GITHUB_USER}/triton-dotfiles.git $HOME/.dotfiles
 else
   git clone https://github.com/${GITHUB_USER}/dotfiles.git $HOME/.dotfiles
 fi
@@ -30,15 +30,15 @@ mkdir -p $HOME/bin
 #export SUDO_ASKPASS=/bin/true
 
 # A bit of a hack
-mv .gitconfig .gitconfig.private
-mv .bashrc .bashrc.dist
-$HOME/.dotfiles/bin/dotfiles.symlink install
+[ -f .gitconfig ] && mv .gitconfig .gitconfig.private  # Added checks
+[ -f .bashrc ] && mv .bashrc .bashrc.dist             # Added checks
+[ -x $HOME/.dotfiles/bin/dotfiles.symlink ] && $HOME/.dotfiles/bin/dotfiles.symlink install || echo "Missing dotfiles.symlink"
 
 # Setting proper terminal
-if [[ -n "$FISH"  ]]; then
+if [[ -n "$FISH"  ]]; then#this
   echo "Setting shell to fish"
   sudo chsh -s /usr/bin/fish $USER
-elif [[ -n "$ZSH" ]]; then
+elif [[ -n "$ZSH" ]]; then#this
   echo "Setting shell to zsh"
   sudo chsh -s /usr/bin/zsh $USER
 else
